refactor: move i18n hooks directory
