+++ src/client/templates/Album/index.tsx
@@ -8,9 +8,9 @@ import { toCdNumber } from 'utils/strings';
 import { TextDivider } from 'client/components/molecules/TextDivider';
 import { MemberCard } from 'client/components/molecules/cards/MemberCard';
 import { getMemberUrl, getSongUrl } from 'client/utils/urls';
-import { useTranslations } from 'client/hooks/useTranslations';
+import { useTranslations } from 'client/i18n/hooks/useTranslations';
 import { HorizontalCard } from 'client/components/molecules/cards/HorizontalCard';
-import { useIntl } from 'client/hooks/useIntl';
+import { useIntl } from 'client/i18n/hooks/useIntl';
 import { GridImage } from 'client/components/atoms/images/GirdImage';
 import { useLanguageContext } from 'client/store/language/hook';
 import { PageHelmet } from 'client/layouts/PageHelmet';
