+++ src/client/templates/Discography/index.tsx
@@ -7,7 +7,7 @@ import { PageContent } from 'client/components/templates/Page';
 import { TextDivider } from 'client/components/molecules/TextDivider';
 import { commonStyles } from 'client/styles/tokens';
 import { getAlbumUrl, getDiscographyUrl } from 'client/utils/urls';
-import { useTranslations } from 'client/hooks/useTranslations';
+import { useTranslations } from 'client/i18n/hooks/useTranslations';
 import { useLanguageContext } from 'client/store/language/hook';
 import { PageHelmet } from 'client/layouts/PageHelmet';
 import { DiscographyPageProps } from 'pages/discography';
