+++ src/client/templates/Member/index.tsx
@@ -10,8 +10,8 @@ import { useLanguageContext } from 'client/store/language/hook';
 import { TextDivider } from 'client/components/molecules/TextDivider';
 import { useAppTheme, commonStyles } from 'client/styles/tokens';
 import { GridMemberImage } from 'client/components/atoms/images/GridMemberImage';
-import { useTranslations } from 'client/hooks/useTranslations';
-import { useIntl } from 'client/hooks/useIntl';
+import { useTranslations } from 'client/i18n/hooks/useTranslations';
+import { useIntl } from 'client/i18n/hooks/useIntl';
 import { PositionCounter } from 'client/templates/Member/components/PositionCounter';
 import { TextLink } from 'client/components/molecules/links/TextLink';
 import { GlowStickColorType } from 'server/actors/Members/constants/glowStickColor';
