+++ src/client/templates/Members/index.tsx
@@ -7,8 +7,8 @@ import { TextDivider } from 'client/components/molecules/TextDivider';
 import { MemberCard } from 'client/components/molecules/cards/MemberCard';
 import { commonStyles } from 'client/styles/tokens';
 import { getMembersUrl, getMemberUrl } from 'client/utils/urls';
-import { useTranslations } from 'client/hooks/useTranslations';
-import { useIntl } from 'client/hooks/useIntl';
+import { useTranslations } from 'client/i18n/hooks/useTranslations';
+import { useIntl } from 'client/i18n/hooks/useIntl';
 import { useLanguageContext } from 'client/store/language/hook';
 import { PageHelmet } from 'client/layouts/PageHelmet';
 import { MembersPageProps } from 'pages/members';
